worldedit branch
The major problem is that WorldEdit v6.0.0 now restricts commands to
players that are online, or seen by Bukkit.getOnlinePlayers();

So that's what this branch is trying to solve.

I use an abstract Injector.inject(vp); Because there is more than one
injection point. It works, but has errors and side-effects.

The major problem with this approach is errors because of the lack of a
PlayerConnection.

https://github.com/TechzoneMC/NPCLib

NPCLib is a possible solution to the lack of a PlayerConnection. We could
add a dependency, and shade in the needed classes, by Techable.

His classes are included, so that I could test them with VirutalPlayers.
But they produced errors, which are included in error.txt

All of this (VirtualPlayers existing in playerList and having a
PlayerConnection) produces a major side-effect... And that is that they
can no longer run the cmd /dc, because they are considered real Players
now, NOT VirtualPlayers.

/dc <virtualplayer> <command> [args]

dc = do command